Remove omnisharp--filter-company-candidate
One of the completions being equal to the prefix is usually normal, and
removing it in the backend means idle completion will always try to
suggest completing to something, even if the current text is valid in
the current context.

If removing completions like this is somehow beneficial, it's probably
better performed in `company-transformers`.