Allow completing from foreign namespaces (off by default).
This can be enabled by setting
omnisharp-auto-complete-want-importable-types to a non-nil value. This
currently only works for ido-mode and popup.el.

Completions from the server will have a tag appended to them (by the
server) like [using System.IO]. These kinds of completions insert the
completion like normally except they also add a using directive at the
very first line in the file. Point is not moved when the directive is
inserted, it's really nice! :)