Add example config for prodigy.el
Prodigy is an emacs plugin that provides an interface within emacs for
launching different kinds of scripts / actions. It also allows
monitoring the state and output of the scripts from within emacs. This
provides a really nice development environment.

The idea behind this config is that developers may take the example
config and adapt it to their own systems.

For prodigy see https://github.com/rejeep/prodigy.el