Implement reading full lines from consecutive partial stdio messages
The omnisharp-roslyn stdio server's output will be sent from the server
one line at a time, but emacs is forced to read it in chunks of some
amount of characters. This will keep track of what characters were last
sent from the server, and return full lines sent from the server.