Allow adding "using System;" via code action started after a symbol
Previously point had to be positioned one column before the end of the
word. This should be more convenient.

After this change, in vanilla emacs, it's impossible to refer to the
current symbol when point is before a word. This might previously been
possible.

Now point has to be inside the start of the word for the server to pick
it up in navigations etc.

In evil-mode, these work:
- evil-normal-state: when point is at the first/last character of a word
- evil-insert-state is the same as vanilla emacs, so works only at end