Revive omnisharp-stop-server and omnisharp-reload-solution
This my attempt at reviving those two interactive commands.

Note that I do not try to issue a command to omnisharp-server to reload
the project (which is probably how it should eventually be) -- I just
kill the old process and restart it with the previously used project
filename/directory name.