Fix eldoc functionality for async server I/O
My last commit 37b22ad5f9af9e26979fbdcaf15f0a1bea89a086 that attempted
to fix eldoc actually did not do that -- what it did is just silenced an
error message.

This refactors omnisharp-eldoc-function to actually do async typelookup
on the symbol under POINT (when omnisharp server is running).

Note this also removes the omnisharp--eldoc-default which attempted to
retrieve information about the function around the code, but I *think*
it is not what eldoc is about. `whichfunc' is the minor mode that is
responsible for that kind of indication.. Thus I gave myself freedom to
remove that functionality.