Fix format of commands sent to stdin server
It appears that the current versions of omnisharp-roslyn require that a command
is prefixed with
`/` (https://github.com/OmniSharp/omnisharp-roslyn/blob/master/src/OmniSharp.Abstractions/OmniSharpEndpoints.cs#L3)
and, indeed, without the prefix present omnisharp-emacs fails on each command
with the `NotSupportedException` exception thrown on each request.

This commit makes sure that the prefix is present.